moved link mem to component, added tw forms
