+++ package.json
@@ -17,6 +17,7 @@
     "@sveltejs/adapter-auto": "^3.0.0",
     "@sveltejs/kit": "^2.0.0",
     "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@tailwindcss/forms": "^0.5.9",
     "@types/luxon": "^3.4.2",
     "@types/node": "22.5.4",
     "autoprefixer": "10.4.20",
