+++ tailwind.config.ts
@@ -1,23 +1,27 @@
-import { join } from 'path'
-import type { Config } from 'tailwindcss'
-import { skeleton } from '@skeletonlabs/tw-plugin'
+import { join } from 'path';
+import type { Config } from 'tailwindcss';
+import { skeleton } from '@skeletonlabs/tw-plugin';
 
 export default {
 	darkMode: 'class',
-	content: ['./src/**/*.{html,js,svelte,ts}', join(require.resolve('@skeletonlabs/skeleton'), '../**/*.{html,js,svelte,ts}')],
+	content: [
+		'./src/**/*.{html,js,svelte,ts}',
+		join(require.resolve('@skeletonlabs/skeleton'), '../**/*.{html,js,svelte,ts}')
+	],
 	theme: {
-		extend: {},
+		extend: {}
 	},
 	plugins: [
+		require('@tailwindcss/forms'),
 		skeleton({
 			themes: {
 				preset: [
 					{
 						name: 'rocket',
-						enhancements: true,
-					},
-				],
-			},
-		}),
-	],
+						enhancements: true
+					}
+				]
+			}
+		})
+	]
 } satisfies Config;
